Feature/#40 redisson lock (#41)
* test: 락 적용 전 테스트

* feat: post like에 락 적용

* feat: comment like에 락 적용

* feat: unlike에 락 적용

* refactor: unlock 예외 처리 리팩토링 및 release 시간 변경

* refactor: 전체 FakeRepository CopyOnWriteArrayList 적용

* refactor: import 정리

+++ src/main/java/com/example/servertech/domain/post/application/PostService.java
@@ -2,6 +2,8 @@ package com.example.servertech.domain.post.application;
 
 import com.example.servertech.common.event.domain.CommonEvent;
 import com.example.servertech.common.event.producer.EventProducer;
+import com.example.servertech.domain.common.entity.exception.LockInterruptedException;
+import com.example.servertech.domain.common.entity.exception.TryLockFailureException;
 import com.example.servertech.domain.post.entity.Post;
 import com.example.servertech.domain.post.entity.PostLike;
 import com.example.servertech.domain.post.exception.NoSuchPostException;
@@ -15,10 +17,13 @@ import com.example.servertech.domain.post.repository.PostRepository;
 import com.example.servertech.domain.user.application.UserService;
 import com.example.servertech.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import static com.example.servertech.common.event.domain.EventType.POST_LIKE;
 import static com.example.servertech.domain.user.entity.UserRole.ADMIN;
@@ -30,6 +35,9 @@ public class PostService {
 	private final PostRepository postRepository;
 	private final PostLikeRepository postLikeRepository;
 	private final EventProducer eventProducer;
+	private final RedissonClient redissonClient;
+
+	private static final String LOCK_KEY_PREFIX = "lock:postlike:";
 
 	@Transactional
 	public PostPersistResponse create(PostCreateRequest request) {
@@ -92,17 +100,42 @@ public class PostService {
 
 	@Transactional
 	public void like(Long id) {
-		Post post = findPostById(id);
-		postLikeRepository.save(
-			PostLike.create(post, userService.me())
-		);
-		eventProducer.publish(
-			CommonEvent.create(post.getWriter().getId(), POST_LIKE)
-		);
+		String lockKey = LOCK_KEY_PREFIX + id;
+		RLock lock = redissonClient.getLock(lockKey);
+		boolean locked = false;
+
+		try {
+			locked = lock.tryLock(1, 30, TimeUnit.SECONDS);
+			if (!locked) throw new TryLockFailureException();
+
+			Post post = findPostById(id);
+			postLikeRepository.save(
+				PostLike.create(post, userService.me())
+			);
+			eventProducer.publish(
+				CommonEvent.create(post.getWriter().getId(), POST_LIKE)
+			);
+		} catch (InterruptedException e) {
+			throw new LockInterruptedException();
+		} finally {
+			if (locked) lock.unlock();
+		}
 	}
 
 	@Transactional
 	public void unlike(Long id) {
-		postLikeRepository.deleteByPostAndUser(id, userService.me().getId());
+		String lockKey = LOCK_KEY_PREFIX + id;
+		RLock lock = redissonClient.getLock(lockKey);
+		boolean locked = false;
+		try {
+			locked = lock.tryLock(1, 30, TimeUnit.SECONDS);
+			if (!locked) throw new TryLockFailureException();
+
+			postLikeRepository.deleteByPostAndUser(id, userService.me().getId());
+		} catch (InterruptedException e) {
+			throw new LockInterruptedException();
+		} finally {
+			if (locked) lock.unlock();
+		}
 	}
 }
